Rename player.html -> display-id.html
