+++ app/display-id.html
@@ -5,7 +5,7 @@
   <link rel="stylesheet" href="css/reset.css" />
   <link rel="stylesheet" href="css/font-awesome.min.css" />
   <link rel="stylesheet" href="css/player.css" />
-  <script src="player.bundle.js"></script>
+  <script src="display-id.bundle.js"></script>
   <title>Rise Vision Player</title>
 </head>
 <body>
