+++ src/display-id.js
@@ -1,5 +1,5 @@
 const validator = require('./display-id-validator');
-const screen = require('./display-id-screen');
+const screen = require('./display-registration');
 
 function init() {
   screen.init(document, validator);
