+++ src/display-registration.js
