+++ webpack.config.js
@@ -2,9 +2,9 @@ const path = require("path");
 
 module.exports = {
   entry: {
-    background: path.join(__dirname, "src", "background.js"),
-    player: path.join(__dirname, "src", "player.js"),
-    viewer: path.join(__dirname, "src", "viewer.js")
+    "background": path.join(__dirname, "src", "background.js"),
+    "display-id": path.join(__dirname, "src", "display-id.js"),
+    "viewer": path.join(__dirname, "src", "viewer.js")
   },
   output: {
     path: path.join(__dirname, "app"),
